chore: remove duplicate field builders
